Move Video predictes and Lecture predicates into appropriate packages
