+++ src/main/java/seedu/address/logic/trackereventsystem/observers/ListObserver.java
@@ -8,12 +8,12 @@ import seedu.address.logic.trackereventsystem.OnVideoEditedEventObserver;
 import seedu.address.model.DisplayListLevel;
 import seedu.address.model.Model;
 import seedu.address.model.lecture.LectureName;
+import seedu.address.model.lecture.LecturePredicate;
 import seedu.address.model.lecture.ReadOnlyLecture;
-import seedu.address.model.module.LecturePredicate;
 import seedu.address.model.module.ModuleCode;
 import seedu.address.model.module.ReadOnlyModule;
-import seedu.address.model.module.VideoPredicate;
 import seedu.address.model.video.Video;
+import seedu.address.model.video.VideoPredicate;
 
 /**
  * Represents a list observer that responds to on module edited and on
