+++ src/test/java/seedu/address/logic/commands/FindCommandTest.java
@@ -21,17 +21,17 @@ import seedu.address.model.Model;
 import seedu.address.model.ModelManager;
 import seedu.address.model.UserPrefs;
 import seedu.address.model.lecture.Lecture;
+import seedu.address.model.lecture.LectureNameContainsKeywordsPredicate;
+import seedu.address.model.lecture.LectureTagContainsKeywordsPredicate;
 import seedu.address.model.lecture.ReadOnlyLecture;
-import seedu.address.model.module.LectureNameContainsKeywordsPredicate;
-import seedu.address.model.module.LectureTagContainsKeywordsPredicate;
 import seedu.address.model.module.Module;
 import seedu.address.model.module.ModuleCode;
 import seedu.address.model.module.ModuleContainsKeywordsPredicate;
 import seedu.address.model.module.ModuleTagContainsKeywordsPredicate;
 import seedu.address.model.module.ReadOnlyModule;
-import seedu.address.model.module.VideoNameContainsKeywordsPredicate;
-import seedu.address.model.module.VideoTagContainsKeywordsPredicate;
 import seedu.address.model.video.Video;
+import seedu.address.model.video.VideoNameContainsKeywordsPredicate;
+import seedu.address.model.video.VideoTagContainsKeywordsPredicate;
 import seedu.address.testutil.LectureBuilder;
 import seedu.address.testutil.ModuleBuilder;
 import seedu.address.testutil.TypicalLectures;
